+++ package.json
@@ -19,7 +19,8 @@
   "dependencies": {},
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
-    "lint": "eslint js/"
+    "lint": "eslint js/",
+    "lint-fix": "eslint --fix js/"
   },
   "engines": {
     "node": "16.x",
